相关搜索模块化
