+++ searcher/model/QuerySim.go
@@ -1,0 +1,6 @@
+package model
+
+type QuerySim struct {
+	QueryId    int
+	Similarity float64
+}
